+++ src/main/java/ru.recutils/common/ObservationHolder.java
@@ -3,9 +3,9 @@ package ru.recutils.common;
 import java.util.Map;
 
 public interface ObservationHolder {
-    Map<Integer, Double> getFeatures();
+    Map<Integer, Float> getFeatures();
 
-    double getLabel();
+    float getLabel();
 
-    double getImportance();
+    float getImportance();
 }
