+++ src/main/java/ru.recutils/io/SimpleObservationHolder.java
@@ -5,26 +5,26 @@ import java.util.Map;
 import ru.recutils.common.ObservationHolder;
 
 public class SimpleObservationHolder implements ObservationHolder {
-    private final Map<Integer, Double> featureWeights;
-    private final double label;
-    private final double importance;
+    private final Map<Integer, Float> featureWeights;
+    private final float label;
+    private final float importance;
 
     @Override
-    public Map<Integer, Double> getFeatures() {
+    public Map<Integer, Float> getFeatures() {
         return featureWeights;
     }
 
     @Override
-    public double getLabel() {
+    public float getLabel() {
         return label;
     }
 
     @Override
-    public double getImportance() {
+    public float getImportance() {
         return importance;
     }
 
-    public SimpleObservationHolder(Map<Integer, Double> featureWeights, double label, double importance) {
+    public SimpleObservationHolder(Map<Integer, Float> featureWeights, float label, float importance) {
         this.featureWeights = featureWeights;
         this.label = label;
         this.importance = importance;
