+++ src/main/java/ru.recutils/lossfuncs/BinaryLogisticLoss.java
@@ -3,18 +3,18 @@ package ru.recutils.lossfuncs;
 public class BinaryLogisticLoss implements LossFunction {
 
     @Override
-    public double value(double prediction, double gt) {
-        double exp = -prediction * gt;
+    public float value(float prediction, float gt) {
+        float exp = -prediction * gt;
         // logsumexp trick
         if (exp < 0) {
-            return Math.log1p(Math.exp(exp));
+            return (float)Math.log1p(Math.exp(exp));
         } else {
-            return exp + Math.log1p(Math.exp(-exp));
+            return exp + (float)Math.log1p(Math.exp(-exp));
         }
     }
 
     @Override
-    public double derivative(double prediction, double gt) {
-        return gt / (1 + Math.exp(gt * prediction));
+    public float derivative(float prediction, float gt) {
+        return gt / (1 + (float)Math.exp(gt * prediction));
     }
 }
